[NOT FOR UPSTREAM] Use code font as terminal's default font
This commit changes the default terminal font to (configured) code font
(`StandardFonts codeFont`) if code font is monospace. If not, fall back
to "DejaVu Sans Mono" of the same size as code font.

Detection whether a font is monospace or not is currently quite simplistic,
by comparing widths of M and i.