Fix closing PTY and terminating I/O reader process
...in case the PTY is externally supplied (for example using 
`TerminalEmulator class >> #openPTY:`). 

Prior this commit, the code simply assumed that 1) PTerm spawns shell
2) and open PTY. This means that I/O reader process terminated when
child process has finished (and has been reaped) and once it finishes,
it closed the PTY. 

But this is not anymore the only use case, one can supply (opened) PTY.
In this case, there's no child (shell) process to wait for and PTerm
should not close the PTY - closing it is responsibility of the caller
(who can subscribe to `WindowClose` event). 

This commit fixes this, using `pid isNil` as test whether or not PTerm
should close the PTY and using `master notNil` as a test whether or not
to terminate the I/O reader process. 

Implementation notice: we use `fd := master.` and then test `fd` since
the value is accessed by multiple threads. This avoid the value to change
between test and appropriate action, exploiting the fact that Pharo is
"massivelly singlethreaded" and assignment is not VM suspension point.